refactoring

+++ server/src/constants.rs
@@ -1,2 +1,2 @@
-pub const STARTING_ELO: usize = 1000;
+pub const STARTING_ELO: u32 = 1000;
 pub const USERS_PER_SET: usize = 10;
